components - comfo: new integration to control Zehnder ComfoAir units
This integrates with the https://github.com/ti-mo/comfo projects.
It's a protobuf/Twirp-based Go API that communicates with the unit
over a serial connection. This is different from the ComfoConnect
cloud solution that requires internet access.

The integration itself is fully async, but the underlying library is not.
The generated Twirp bindings in the `comfo` PyPI package are still
python2-ish and don't implement async http (yet).

The integration is released with the following features:
- config flow
- a Fan for controlling the unit's fan duty level
- Sensors for the unit's temperature and duty readings
- Binary Sensor(s) for the unit's error states (replace filter, etc.)
- Services for modifying the unit's fan duty speeds/presets